requirements.txt
langchain==0.1.16
langchain-google-genai==0.0.9
langgraph==0.0.40
google-generativeai==0.3.2
pinecone-client==3.0.0
streamlit==1.31.0
pypdf==4.0.1
python-dotenv==1.0.1
